+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 
 
+## Unreleased
+
+### Changed
+
+- Use go-kit's logger interface
+
+### Removed
+
+- Logrus hooks
+
+
 ## 0.1.1 - 2017-06-19
 
 ### Added